moved logic to fetch fallback keys, identifiy displayable key and use of key names cache into dedicated classes
